requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pytest

# RAG / Gemini (NLP_Capstone_project)
langchain
langchain-google-genai
chromadb
langchain-chroma
google-generativeai

# ETL related libraries
pandas
numpy
textblob